PDFBOX-4497: remove cast to keep it a float
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/1.8@1856430 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdfviewer/PageDrawer.java
@@ -428,7 +428,7 @@ public class PageDrawer extends PDFStreamEngine
                     float w = transformWidth(dashArray[i]);
                     dashArray[i] = Math.max(w, 0.062f);
                 }
-                phaseStart = (int)transformWidth(phaseStart);
+                phaseStart = transformWidth(phaseStart);
 
                 // empty dash array is illegal
                 // avoid also infinite and NaN values (PDFBOX-3360)
